Update README.md
# irrigation
a working irrigation system using Arduino uno


During the process of manual irrigation, there are situations where the right amount of water is not provided to plants or more than necessary water is provided. Both cases pose threats to plants and their growth
The proposed system is based on the parameters like soil moisture, water level in tank/ storage container and atmospheric temperature. Input from the given parameters, determine the conditions that need to be satisfied for irrigation. With Arduino Uno board as the brain, the system receives information about the soil moisture level using a Soil Moisture Sensor,